// src/generated/models/movie.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * Cinema Backend API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Фильм
 */
export interface Movie {
  /** Идентификатор фильма */
  id?: number;
  /** Название фильма */
  title?: string;
  /** Описание фильма */
  description?: string;
  /** Год выпуска */
  year?: number;
  /** Продолжительность в минутах */
  lengthMinutes?: number;
  /** URL постера */
  posterImage?: string;
  /** Рейтинг */
  rating?: number;
}